(fix) Added Yazio in MCP tool
