+++ SparkyFitnessServer/services/externalFoodSearchService.ts
@@ -25,22 +25,13 @@ import {
   searchNorishFoods,
 } from './foodIntegrationService.js';
 
-export const VALID_PROVIDER_TYPES = [
-  'openfoodfacts',
-  'usda',
-  'fatsecret',
-  'mealie',
-  'tandoor',
-  'yazio',
-  'norish',
-  'swissfood',
-] as const;
+import type { ProviderType } from '../constants/foodProviders.js';
 
-export type ProviderType = (typeof VALID_PROVIDER_TYPES)[number];
-
-export function isValidProviderType(value: string): value is ProviderType {
-  return (VALID_PROVIDER_TYPES as readonly string[]).includes(value);
-}
+export {
+  VALID_PROVIDER_TYPES,
+  isValidProviderType,
+} from '../constants/foodProviders.js';
+export type { ProviderType } from '../constants/foodProviders.js';
 
 export interface ProviderCredentials {
   app_id?: string;
